Make vault_container_version fixture available
The import from the support module was missed in the migration. Since
we're now dedicated to Vault, we can just put it into the root conftest.
